chore(nexus): nexus child retire logic cleanup
